Moved index.php to a pure html/js index.html
Moved all layout logic out of index.php into javascript and a
renderedlayouts.php file. This is a first step in moving off of PHP to hopefully
a full javascript stack to allow for flexibility in the future. Additionally
there is a layouts.php file which will return a json object containing all the
layouts that exist. The intent is for renderedlayouts.php to be temporary and be
replaced with client side templating.