Somewhat fix overlapping layer labels beyond layer 2
Previously layers 1, 3, 5 and 7 were positioned to the top
center, layers 2, 4 and 6 to the bottom center. These labels
were unreadable in all cases, since they overlapped each other.

This change places the labels like so:

+-------+
| 3 1 5 |
|   0 7 |
| 4 2 6 |
+-------+

The positions are chosen to maintain backwards compatibility
with the old positioning with default layers 0-2.

Of course there will still always be overlap with wider labels.