Create WhiteFox Specific Diff Algorithm
The current kll diff algorithms assume the same number of keys and order as the
defaultMap.kll deep in the controller code. With the WhiteFox this is
problematic as the different layouts all have different numbers of keys. These
changes add a new WhiteFox specific algorithm to detect changes based on the
scan codes rather than order so we can have a Base that has more keys than our
current map which will be ignored.