Temporary fix for 60% LCD Alphabet layout
This should fix the issue with the swapped Backspace and Backslash keys on the
Alphabet layout. NOTE: This is a bit ugly since I have to swap the keys on the
Blank version, but it should suffice until I move all keyboards over to the new
KLL generation mechanism used for the WhiteFox.